+++ packages/playground/dynamic-import/nested/index.js
@@ -7,9 +7,17 @@ async function setView(view) {
   document.querySelector(`.${id}`).addEventListener('click', () => setView(id))
 })
 
+// literal dynamic
 document.querySelector('.baz').addEventListener('click', async () => {
-  // literal dynamic
   const { msg } = await import('../views/baz.js')
+  text('.view', msg)
+})
+
+// full dynamic
+const arr = ['qux.js']
+const view = `/${arr[0]}`
+document.querySelector('.qux').addEventListener('click', async () => {
+  const { msg } = await import(/*@vite-ignore*/ view)
   text('.view', msg)
 })