+++ packages/playground/dynamic-import/views/foo.js
@@ -1,1 +1,4 @@
+import { n } from '../nested/shared'
+console.log('foo' + n)
+
 export const msg = 'Foo view'